Move Indentation, Cursor, Selection out of the `editor` class
Background: I am currently working on migrating
[Notepadqq from Makefiles to Bazel](https://github.com/Danaozhong/notepadqq/pull/1).
As a part of this process, I am splitting up the code into individual
libraries. This significantly decreases compile time and improves
modularity.
Right now, the `notepadqq` binary is one big monolith, where all
files are compiles together. This manifests in the need for a lot
of forward declaractions in header files to break up circular
dependencies.
Splitting this into libraries requires properly resolving circular
dependencies, and I plan to merge these into the main codebase
one change at a time.

In this PR, I am resolving a circular dependency between the `editor`
class and the `Search` and `Sessions` code: right now, the editor
includes the `Search` and `Sessions` headers, and their implementation
include the `editor` again.
I resolved this circular dependency by moving the `Indentation`,
`Cursor` and `Selection` into a separate file, that can be included
by both `editor` and `Search`/`Sessions`.

This change was tested locally by building the application, and
running it.
