+++ src/ui/include/EditorNS/editor.h
@@ -3,6 +3,7 @@
 
 #include "include/EditorNS/customqwebview.h"
 #include "include/EditorNS/languageservice.h"
+#include "include/EditorNS/editor_properties.h"
 
 #include <QObject>
 #include <QQueue>
@@ -93,40 +94,7 @@ namespace EditorNS
 
         static void invalidateEditorBuffer();
 
-        struct Cursor {
-            int line;
-            int column;
-
-            bool operator == (const Cursor &x) const {
-                return line == x.line && column == x.column;
-            }
-
-            bool operator < (const Cursor &x) const {
-                return std::tie(line, column) < std::tie(x.line, x.column);
-            }
-
-            bool operator <= (const Cursor &x) const {
-                return *this == x || *this < x;
-            }
-
-            bool operator > (const Cursor &x) const {
-                return !(*this <= x);
-            }
-
-            bool operator >= (const Cursor &x) const {
-                return !(*this < x);
-            }
-        };
-
-        struct Selection {
-            Cursor from;
-            Cursor to;
-        };
-
-        struct IndentationMode {
-            bool useTabs;
-            int size;
-        };
+        
 
         /**
          * @brief Just a flag that is used for marking editors that are still loading,
@@ -312,9 +280,10 @@ namespace EditorNS
          * @return a pair whose first element is the document indentation, that is
          *         significative only if the second element ("found") is true.
          */
-        QPromise<std::pair<IndentationMode, bool>> detectDocumentIndentation();
-        Editor::IndentationMode indentationMode();
-        QPromise<IndentationMode> indentationModeP();
+        QtPromise::QPromise<std::pair<IndentationMode, bool>> detectDocumentIndentation();
+        IndentationMode indentationMode();
+        QtPromise::QPromise<IndentationMode> indentationModeP();
+
 
         QPromise<QString> getCurrentWord();
